+++ packages/annotation-converter/src/converter.ts
@@ -765,19 +765,22 @@ function parseCollection(
 
         case 'NavigationPropertyPath':
             return collectionDefinition.map((navPropertyPath, navPropIdx) => {
+                const navigationPropertyPath = navPropertyPath.NavigationPropertyPath ?? '';
                 const result = {
                     type: 'NavigationPropertyPath',
-                    value: navPropertyPath.NavigationPropertyPath,
+                    value: navigationPropertyPath,
                     fullyQualifiedName: `${parentFQN}/${navPropIdx}`
                 } as any;
 
-                lazy(
-                    result,
-                    '$target',
-                    () =>
-                        resolveTarget(converter, currentTarget, navPropertyPath.NavigationPropertyPath, currentTerm)
-                            .target
-                );
+                if (navigationPropertyPath === '') {
+                    result.$target = undefined;
+                } else {
+                    lazy(
+                        result,
+                        '$target',
+                        () => resolveTarget(converter, currentTarget, navigationPropertyPath, currentTerm).target
+                    );
+                }
 
                 return result;
             });
@@ -1036,7 +1039,11 @@ class Converter {
     }
 
     getConvertedActionImport(fullyQualifiedName: FullyQualifiedName) {
-        return this.convertedOutput.actionImports.by_fullyQualifiedName(fullyQualifiedName);
+        let actionImport = this.convertedOutput.actionImports.by_fullyQualifiedName(fullyQualifiedName);
+        if (!actionImport) {
+            actionImport = this.convertedOutput.actionImports.by_name(fullyQualifiedName);
+        }
+        return actionImport;
     }
 
     getConvertedAction(fullyQualifiedName: FullyQualifiedName) {