+++ packages/fe-mockserver-core/src/mockdata/fileBasedMockData.ts
@@ -330,6 +330,11 @@ export class FileBasedMockData {
                 );
             }
         });
+        this._entityType.navigationProperties.forEach((navigationProperty) => {
+            if (navigationProperty.containsTarget) {
+                outObj[navigationProperty.name] = [];
+            }
+        });
 
         return outObj;
     }
@@ -379,7 +384,7 @@ export class FileBasedMockData {
         actionData: any,
         _keys: Record<string, any>,
         _odataRequest: ODataRequest
-    ): Promise<object> {
+    ): Promise<object | undefined> {
         return actionData;
     }